Use a single function for generating CSVs
